Move auth to layers/auth
